"5:50pm.
```
Starting timing for: iteration 0
The time was 00:00:07.4330474 for: iteration 0
Training: 0.523186318914637
Testing: 0.305616292375362(9568/9984)
Starting timing for: iteration 1
The time was 00:00:07.3323087 for: iteration 1
Training: 0.193773912839936
Testing: 0.239841713572731(9647/9984)
Starting timing for: iteration 2
The time was 00:00:07.3254822 for: iteration 2
Training: 0.0784469467260314
Testing: 0.208978438993276(9697/9984)
Starting timing for: iteration 3
The time was 00:00:07.3455555 for: iteration 3
Training: 0.0397165229915117
Testing: 0.204663403217474(9714/9984)
Starting timing for: iteration 4
The time was 00:00:07.3455595 for: iteration 4
Training: 0.0213219665741274
Testing: 0.171711973308137(9765/9984)
Starting timing for: iteration 5
The time was 00:00:07.3344142 for: iteration 5
Training: 0.0116269622897363
Testing: 0.159398061164149(9785/9984)
Starting timing for: iteration 6
The time was 00:00:07.3289676 for: iteration 6
Training: 0.00580885296947258
Testing: 0.140541458438642(9799/9984)
Starting timing for: iteration 7
The time was 00:00:07.3278100 for: iteration 7
Training: 0.00230095739067492
Testing: 0.156616333721865(9786/9984)
Starting timing for: iteration 8
The time was 00:00:07.3471488 for: iteration 8
Training: 0.000471746435499055
Testing: 0.138228890517832(9808/9984)
Starting timing for: iteration 9
The time was 00:00:07.3510650 for: iteration 9
Training: 0.000122740214832216
Testing: 0.138717759758735(9799/9984)
Starting timing for: iteration 10
The time was 00:00:07.3420016 for: iteration 10
Training: 2.37564564178889E-05
Testing: 0.138499770681472(9804/9984)
Starting timing for: iteration 11
The time was 00:00:07.3366106 for: iteration 11
Training: 2.22116714330126E-05
Testing: 0.138077182659457(9806/9984)
Starting timing for: iteration 12
The time was 00:00:07.3322608 for: iteration 12
Training: 1.64431706374988E-05
Testing: 0.137483125160713(9805/9984)
Starting timing for: iteration 13
The time was 00:00:07.3330441 for: iteration 13
Training: 1.33098594390558E-05
Testing: 0.137348527318243(9805/9984)
Starting timing for: iteration 14
The time was 00:00:07.3331201 for: iteration 14
Training: 1.25055694454265E-05
Testing: 0.137138361297906(9807/9984)
```

This is ln_relu with ...

```
inl learning_rate = epsilon -8
inl pars = {rate={weight=learning_rate; covariance=learning_rate ** 0.85f32; noise=epsilon -3}}
Console.writeline pars
```

Seriously, I will just print out the damn parameters from here. Let me set noise to 0 and remove that sqrt. It does not seem like high noise is a virtue. It actually leeds to the net diverging and overfiting in strange ways.

```
{rate = {covariance = 0.008974205; noise = 1; weight = 0.00390625}}
Starting timing for: iteration 0
The time was 00:00:07.4242479 for: iteration 0
Training: 0.516270745669006
Testing: 0.300121142550782(9570/9984)
Starting timing for: iteration 1
The time was 00:00:07.3597523 for: iteration 1
Training: 0.177150565387392
Testing: 0.231808447288057(9633/9984)
Starting timing for: iteration 2
The time was 00:00:07.3988231 for: iteration 2
Training: 0.0760107061893621
Testing: 0.193126630484293(9706/9984)
Starting timing for: iteration 3
The time was 00:00:07.3647248 for: iteration 3
Training: 0.0359019309562716
Testing: 0.174414718746045(9748/9984)
Starting timing for: iteration 4
The time was 00:00:07.4360277 for: iteration 4
Training: 0.0182920961027493
Testing: 0.16210165633653(9760/9984)
Starting timing for: iteration 5
The time was 00:00:07.4193195 for: iteration 5
Training: 0.00963893945632753
Testing: 0.155456966114819(9793/9984)
Starting timing for: iteration 6
The time was 00:00:07.4236404 for: iteration 6
Training: 0.00396781125342511
Testing: 0.140791446886706(9795/9984)
Starting timing for: iteration 7
The time was 00:00:07.4459549 for: iteration 7
Training: 0.0011413940232127
Testing: 0.139295514590051(9795/9984)
Starting timing for: iteration 8
The time was 00:00:07.4164459 for: iteration 8
Training: 0.000213687247542848
Testing: 0.137647111765546(9796/9984)
Starting timing for: iteration 9
The time was 00:00:07.3480144 for: iteration 9
Training: 1.97769310253673E-05
Testing: 0.137071603374299(9797/9984)
Starting timing for: iteration 10
The time was 00:00:07.3509370 for: iteration 10
Training: 1.42279660053304E-05
Testing: 0.136569474366819(9797/9984)
Starting timing for: iteration 11
The time was 00:00:07.3476048 for: iteration 11
Training: 1.22663973107366E-05
Testing: 0.136201721757816(9796/9984)
Starting timing for: iteration 12
The time was 00:00:07.3505510 for: iteration 12
Training: 1.09714369837689E-05
Testing: 0.135869060060571(9795/9984)
Starting timing for: iteration 13
The time was 00:00:07.3283220 for: iteration 13
Training: 9.7729649459413E-06
Testing: 0.135546384998626(9796/9984)
Starting timing for: iteration 14
The time was 00:00:07.3495816 for: iteration 14
Training: 9.24933189174236E-06
Testing: 0.135303295358105(9796/9984)
```

I think noise 0 should make for a decent default. Let me try setting it to 1. Then I'll try tanh again.

Yeah, it seems -3 was roughly the same noise level as what I have now without it at 0 and without sqrt.

```
{rate = {covariance = 0.008974205; noise = 2; weight = 0.00390625}}
Starting timing for: iteration 0
The time was 00:00:07.4207330 for: iteration 0
Training: 0.505198355333314
Testing: 0.305693407036298(9520/9984)
Starting timing for: iteration 1
The time was 00:00:07.3497516 for: iteration 1
Training: 0.176516252156695
Testing: 0.233874106437955(9655/9984)
Starting timing for: iteration 2
The time was 00:00:07.3671249 for: iteration 2
Training: 0.0739021315434037
Testing: 0.20247492470338(9715/9984)
Starting timing for: iteration 3
The time was 00:00:07.3932064 for: iteration 3
Training: 0.0367268062483244
Testing: 0.182928551350595(9752/9984)
Starting timing for: iteration 4
The time was 00:00:07.3745901 for: iteration 4
Training: 0.018250371149073
Testing: 0.170301080984489(9761/9984)
Starting timing for: iteration 5
The time was 00:00:07.3585950 for: iteration 5
Training: 0.00761863418045364
Testing: 0.177201881687849(9758/9984)
Starting timing for: iteration 6
The time was 00:00:07.3621449 for: iteration 6
Training: 0.0040263471565903
Testing: 0.146450539206925(9784/9984)
Starting timing for: iteration 7
The time was 00:00:07.3287218 for: iteration 7
Training: 0.00174788739139883
Testing: 0.158671317196392(9780/9984)
Starting timing for: iteration 8
The time was 00:00:07.3264164 for: iteration 8
Training: 0.000412402268503109
Testing: 0.15595052363251(9792/9984)
Starting timing for: iteration 9
The time was 00:00:07.3291061 for: iteration 9
Training: 2.56317410016312E-05
Testing: 0.155571453175028(9791/9984)
Starting timing for: iteration 10
The time was 00:00:07.3318191 for: iteration 10
Training: 1.57813032717905E-05
Testing: 0.155084075017557(9795/9984)
Starting timing for: iteration 11
The time was 00:00:07.3582224 for: iteration 11
Training: 1.29012845412914E-05
Testing: 0.15470064158546(9795/9984)
Starting timing for: iteration 12
The time was 00:00:07.3353245 for: iteration 12
Training: 1.18916034349728E-05
Testing: 0.154375824083214(9795/9984)
Starting timing for: iteration 13
The time was 00:00:07.3311804 for: iteration 13
Training: 1.1199296114104E-05
Testing: 0.154013534615095(9795/9984)
Starting timing for: iteration 14
The time was 00:00:07.3326623 for: iteration 14
Training: 9.73440921498961E-06
Testing: 0.153833960474685(9795/9984)
```

Let me do `tanh`. Then I'll do `ln_relu` again. After that I'll call it a day. Weight sampling is not that bad. It is not spectacular, but it is not bad. It should benefit me in RL."