https://www.reddit.com/r/reinforcementlearning/comments/9z2b0t/r_practical_bayesian_learning_of_neural_networks/eauizu0/?st=jp5d19qq&sh=063132c3
"So far I've tested KFAC + weight sampling on Mnist, but there is not much difference in the SL regime. I guess this is why all the papers that I've seen use Bayesian optimization for pruning or uncertainty estimation rather than generalization.

As it turns out, since I posted this paper I did find that my idea has been tried before such as in the [noisy natural gradient paper](https://arxiv.org/abs/1712.02390) and the [scalable laplace paper](https://openreview.net/forum?id=Skdvd2xAZ). The first paper in particular uses KFAC for exploration in a RL domain and finds that it strongly outperforming the alternatives. The exploration method though [Vime](https://arxiv.org/abs/1605.09674) which uses a dynamics model rather than the thing I came up with.

A week ago I thought the Badam paper I posted was unique, but the noisy NG paper has an almost identical derivation of noisy Adam to it. So it is not some lucky break like I'd imagined it to be, but a scenario of different groups converging on similar ideas simultaneously.

Personally, when I tried this yesterday on Mnist I was disappointed that I could not observe a powerful regularization effect - as unlikely as it was, I'd hoped it would shoot straight to 99% on the first epoch given the vaunted generalization benefits of BNNs, but no such luck.

Nonetheless, I think going forward with this for RL and disregarding the SL results should be a good idea. Toy tasks unrelated to the ones of interest can be misleading. I do not see an evolutionary path leading to mastering RL that does not involve making ANNs more Bayesian.

Also as methods like Vime are so complex, I think my idea inspired by random distillation and bootstrapping would be worth trying out. There should definitely be something simpler than Vime out there."