https://youtu.be/h6ajHkgM3ZU?list=PLe5rNUydzV9Q01vWCP9BV7NhJG3j7mz62 [DeepBayes2018]: Day 1, practical session 2. Bayesian reasoning
Right now I'll go through some of these videos. I am thinking that since I am pretty good at programming, the best way to get better at ML long term is to get familiar with everything Bayesian, otherwise I will have to rely on unreliable stabs in the dark to make anything work well. I think that given a few years, I can master everything that I would need in ML to get myself up to the level of where I am now in programming.

It is only in the last six months that I've been able to do serious ML work. In 2016 prior to starting work on Spiral, I did do it for a year, but that was still mostly just building a ML library in F#.

I am going to need a high level of ML skill for the sake of the greater plan.

Yesterday it was like I fired a spell that left me completely exhausted even now. I really want to stop programming for a while until I recharge and the best way to spend that time is on education. Moving between study and programming is a good strategy. The pace I have now is good as I'll improve both my ML tooling and ML knowledge in tandem thanks to Spiral.

I think tomorrow, I'll aim to get the char RNN back online and then I'll get to work on poker. I'll put in entropy regularization, L2 regularization, intrinsic motivation and whatever else is needed to get absolutely the most on the toy task.

Then I'll upscale the task to two cards and then leduc poker after that. Slowly I'll move closer and closer to the goal.

The way to succeed at anything in programming is to cut the big task into numerious small ones and do them one at a time. That has been my strategy up to now. Without a doubt, by the end of the month I will have something tangible.

For now, let me just squeeze out the langour out of my system. I'll begin work when I start feeling up to it.