While GS is downloading, let me put the last few words in for the day.
I keep thinking about the intrinsic motivation scheme I made up and I think it really does have a decent shot of working.

Bootstrapped value heads are not measuring the uncertainty of the value which would be the scale if the value was represented by a gaussian. Instead they measure the meta-uncertainty of the value.

If the mean of the value was represented by a gaussian then that meta-uncertainty would be the scale of that. That scale represents my confusion. It is the scale of the mean. By taking the difference of confusions at the two state I get back the intrinsic reward. It is simple and easy and I do not need a model nor excessively complicated optimization process to get it out.

It will work, it has to work.

So weight sampling is not quite the super strong regularizer that I expected it would be - it is not a problem.

The logic of bootstraping still stands and that is what trully matters.

I'll crack intrinsic motivation and crack poker. Along the way I'll gradually master Bayesian inference and figure out how to get to the next level with differentiable programming.

What exists now cannot possibly be everything. As scalable as backprop is, it is not human brain level scalable. There should be more to it and it will involve messing with gradients. There is an untapped goldmine there.

KFAC + weight sampling cannot possibly the last word in how to optimize neural nets, that I am certain of.

Apart from that, doing model based RL by naively predicting the next state strikes me as foolish. There should be more to it somehow.

I am really inspired at how LSTD does credit assignment. There should be more that is possible to do with optimizers.

Well, I'll stop here for the day. It is true that I did not do much, but I am mostly recovering from yesterday's possession. I took the chance to watch a few lectures, but honestly Bayesian learning is boring me to death.

I'll have to continue it tomorrow. I should aim to go through all the lectures in the next week.

Tomorrow I'll split between programming and studying.

I'll get the char RNN and the second task operational once more. Then comes RL. In a week I'll have the first iteration of the perfected agent.