It is really quite different than when I'd taken the probabilistic graphical models course in 2015. Bayesian modeling is quite simpler than than the algorithms that do inference - or rather, tackling modeling in isolation to inference is surely a best practice.
Probably most of my difficulties in the past to internalize this material is due to the blurring of the two stages.

Model based ML book and this one are definitely a move in the right direction.

I viable approach towards mastering probabilistic programming would be to implement the examples in the ProbMods book along with the inference scheme on my own. That would definitely be doable as the examples are strongly motivated - the PGM course in comparison just drops you off in some desert without a compass.

Right now I am done with chapter 12. I was strongly focused up to 2pm or so and after that I was mostly skimming due to running out of steam.

Chapter 12 is particularly interesting.

Yes, with multiple layers you would expect to get transfer learning automatically - this is something that is not at all the case in deep learning. DL requires special optimizers or RNNs to get that sort of effect.

Blessing of abstraction is definitely a good term to keep in mind. I like how it goes with my handle.

All in all, the book is too hard to go through in one day. I'll finish it tomorrow and resume programming on Tuesday.

Mastering Bayesian ML will be a long term project, I cannot let it distract me at this juncture. I have to see how much I can push RNNs + weight sampling in RL. After I am done with that I'll have leeway to play with this.

The end goal of all of this will be to unify Bayesian inference with gradient based optimization. That will result in scalable, living intelligence in silico.