I got up even later than yesterday. Damn. Well, let me continue the study session. The talk and the PhD thesis remains. I want to read the papers by Wood after that.
When I am done, I'll genuinely start programming. It is kind of ironic that I'd have to give up on the current approach to continue with it, but I'll bear with it for now. Only good things can come from continuing the course and not giving up...not giving up all the way that is.