Let me call it a day with this.
https://www.reddit.com/r/ProgrammingLanguages/comments/a3euun/frank_wood_probabilistic_programming_and/

"Frank Wood's approach to probabilistic programming which involves using deep nets as inference accelerators is rather new to me and I found this talk which is evenly split between PL and ML to be quite interesting. There is actually a wide variety of approaches to probabilistic programming, but I think that this particular one has a lot of potential for synergy as it allows exact inference at high speed. Typically, for more naive approaches they either do approximate (variational) inference which is less accurate or much slower Monte Carlo-styled inference which is exact.

The experience of last six months has convinced me that deep learning by itself won't really be enough to cross the huge gap in sample complexity needed to make it as efficient in learning as humans. The Bayesian framework in general and more idealized Bayesian models have an answer for how to bridge it and do optimal learning, but have been held back by scalability issues for a long time.

As a resolution for the next few years, rather than make stabs in the dark like I have so far, I'll make getting good at probabilistic programming my goal. Things like learning at rates humans do is something that is only possible with very strong priors, and good exploration which is vital to RL is possible only if the agent has the capability to estimate uncertainty. Deep learning can do a little bit of that using noise injection, but if a little bit is good then one might as well go the whole way and do it properly.

As an aside, I really wish this talk was in higher resolution as the slides are hard to read. They are partly made from [his talks](https://www.cs.ubc.ca/~fwood/talks/index.html) in 2015, so that might help."