"2pm. Damn. I was no counting on this. I mean the error due to state loading. Agh...
Let me bring the neutral element back in.

```
                inl num_valid = near_to % by |> inl x -> if x = 0 then by else x
                inl d =
                    match d with
                    | {finally} -> {d with finally=finally num_valid}
                forcd {d with from=from+by; state=state (index_convert from); by near_to
                    body=inl d -> self {d with i = index_convert self}
                    }
```

No actually, let me first think a bit more carefully. Why is the error even happening?

Remind me again...where is the finally executed?

```
    whilecd {
        state
        cond = inl {from state} -> check from
        body = inl {from state} -> {state=body {state i=from}; from=from+by}
        } .state
    |> finally
```

After the loop.

Hmmm...what would happen if I did that explicitly?

2:15pm. I get a type error.

```
            if from < near_to then
                forcd {d with from=from+by; state=state (index_convert from); by near_to
                    body=inl d -> self {d with i = index_convert self}
                    }
```

The reason is this if statement. I am sure of it, the reason why I am getting errors is not because some threads act out of bounds, but rather because the block reduction happens in the wrong scope. If some threads are inactive then the block reduce will reduce in undefined behavior.

...Though, this is strange. Actually, that can't be right.

No, the reason can't possibly be threads being inactive.

...How about I try running without the if statement. I want to see if the error still happens if I do that."