I understand what the problem is. Just not how to solve it.
```
        | .state_loading {d with state} ->
            assert (by < span dim) "The step size must be less than the dimension size."
            forcd {d with from=from+by; state=state (index_convert from); by near_to
                body=inl d -> self {d with i = index_convert self}
                }
```

Like this? But once I have this I can get rid of that state loading thing and get it directly from the body instead. That would be an improvement.

For the time being, let me just think a little of how I want to make sure that the `span` is less than `by` in the kernel.