What a mess this turned out to be. At least I am done with it.
This is one of the disadvantages of doing it all on your own. Sudden bugs can spring up and take significant time.

I was going to domething else today apart from this thing, but I forgot what it was...

Ah, yes, that would be the char RNN. I think I could give it a shot. There is still some time left in the day.

Let me do it.