"2:30pm.
```
Error trace on line: 263, column: 23 in file "CudaAD".
                    | {primal adjoint} -> primal .get |> dr
                      ^
Error trace on line: 264, column: 23 in file "CudaAD".
                    | x -> x.get
                      ^
Cannot find a member named get inside the module.
```

What do I do about this?

Let me figure out at what point in the test is it occurring.

2:35pm.

```
    inl ty, {run truncate} =
        Union.infer {
            run={
                map=inl {network s} {input label} ->
                    inl network, input = run s input network
                    inl {out bck} = final label input s
                    inl out _ = s.CudaTensor.get out |> to float64
                    inl prev = {out bck=Struct.map (inl {bck} -> bck) network, bck}
                    inl network = Struct.map (inl x -> {x without bck}) network
                    {network s prev}
                input=const {input=input (dyn 0) (dyn 0); label=label (dyn 0) (dyn 0)}
                block=()
                }
```

Hmmm...it is not in here. Then it should be in truncate."