"2:35pm.
```
    inl ty, {run truncate} =
        Union.infer {
            run={
                map=inl {network s} {input label} ->
                    inl network, input = run s input network
                    inl {out bck} = final label input s
                    inl out _ = s.CudaTensor.get out |> to float64
                    inl prev = {out bck=Struct.map (inl {bck} -> bck) network, bck}
                    inl network = Struct.map (inl x -> {x without bck}) network
                    {network s prev}
                input=const {input=input (dyn 0) (dyn 0); label=label (dyn 0) (dyn 0)}
                block=()
                }
```

Hmmm...it is not in here. Then it should be in truncate.

```
inl network, input = run s input network
```

Ok, it is in here. The error is so deep and the error reporting is so bad, that I have no choice but to do it by hand. Spiral's refactoring capabilities suck. I need to make sure that I do not bring myself into this position again. Hopefully this will be the last time that I need to do this.

2:40pm. Strangely I got downvoted by somebody in that thread.

Forget that. Focus. I now know for sure that the type error is in run. Let me go all the way through."