"3:50pm. Focus.
```
                    print_static "qwe"
                    inm data = concat {bias=one; input; state=out}
                    print_static "asd"
```
...
```
lit qwe
lit asd
lit qwe
Error trace on line: 635, column: 21 in file "Learning".
                    |> inl input -> apply input s |> stack
```

So the error is here and happens on the second iteration.

```
type (float32)
type (float32)
lit asd
type (float32)
{adjoint=type (float32); block=[]; primal=type (float32)}
```

I get the sense that the function does not like the state being a dual.

So this is what I need to look into then.

```
{elem_type=type (float32)}
{elem_type={adjoint=type (float32); block=[]; primal=type (float32)}}
{elem_type={adjoint=type (float32); block=[]; primal=type (float32)}}
```

Hmmm, it does feel like it is inferring the type properly.

```
lit qwe
{elem_type=type (float32)}
lit asd
{elem_type={adjoint=type (float32); block=[]; primal=type (float32)}}
lit qwe
{elem_type={adjoint=type (float32); block=[]; primal=type (float32)}}
```

So this middle one is in the linear layer. Yeah, can't forget that one.

```
| {primal adjoint}, _ -> {primal=x; block=()}
```

Is this one really right?

```
{elem_type=type (float32); x=<function>}
{ar=<function>; offset=var (int64); size=[lit 128i64, lit 1i64]}
{elem_type=type (float32); x=<function>}
{ar=<function>; offset=lit 0i64; size=[lit 128i64, lit 1i64]}
{elem_type={adjoint=type (float32); block=[]; primal=type (float32)}; x={adjoint=<function>; block=[]; primal=<function>}}
```

```
Error trace on line: 192, column: 13 in file "Learning".
            print_static x.elem_type
            ^
Cannot find a member named elem_type inside the module.
```

Huh, what the hell is this?

4:05pm. Yeah, since I am doing this in update body, it is no wonder that I am getting this sort of error.

...It feels to me that there is nothing wrong with `concat`. Let me put the print statement where the error actually occurs. This is just so strange. If needed I'll put in some print statements in `load`. Why am I not getting a type error there?

```
Error trace on line: 90, column: 21 in file "CudaFun".
                inl o = init i
                    ^
Error trace on line: 91, column: 17 in file "CudaFun".
                out .view i .set o
                ^
Expected a function in function application.
```"