"5:45pm.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:10.1544471 for: iteration 0
Training: 2.91770538612378
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:10.0959557 for: iteration 1
Training: 2.80988668592866
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:10.1687423 for: iteration 2
Training: 3.86636574439677
The learning rate is 2 ** -13
Starting timing for: iteration 3
```

So this is the power of weight sampling?

Agh, I am not sure I feel like debugging it, but let me turn off weight sampling for just a bit.

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:06.9965140 for: iteration 0
Training: 2.05715640552808
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:06.9329873 for: iteration 1
Training: 1.7084574932164
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:06.9306856 for: iteration 2
Training: 1.63791715574232
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:06.9339008 for: iteration 3
Training: 1.60060050264891
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:06.9462407 for: iteration 4
Training: 1.57717967781526
```

I think I might have made it so that it samples every time by accident somehow. But what is this difference in optimization and this sudden slowness?

Every time I do a refactor it seems something greatly changes.

```
{rate = {covariance = 0.0001451669; weight = 3.051758E-05}}
The learning rate is 2 ** -15
Starting timing for: iteration 0
The time was 00:00:10.1914881 for: iteration 0
Training: 3.13041727240745
The learning rate is 2 ** -15
Starting timing for: iteration 1
The time was 00:00:10.0618356 for: iteration 1
Training: 2.66248432612594
The learning rate is 2 ** -15
Starting timing for: iteration 2
The time was 00:00:10.0826601 for: iteration 2
Training: 3.11621130331803
The learning rate is 2 ** -15
Starting timing for: iteration 3
The time was 00:00:10.1133878 for: iteration 3
Training: 4.24560672407011
The learning rate is 2 ** -15
Starting timing for: iteration 4
```

It is just so slow and unstable now. And it takes nearly 4s from compile up from 2.6s in the previous iteration of the library which was already too slow as it was.

This is the pits.

Just why is it 3s slower with weight sampling turned on? My first suspicion is that I am sampling on every step, but that can't be it.

Who knows.

All I know is that it is nearly 6pm, and I do not feel like continuing this stuff. I am definitely taken aback at how hard it was to get the char RNN to even compile.

```
    inl weight_sample {d with stddev weight front back} s =
        match s.data with
        | {rate} ->
            inl random = s.CudaRandom.create {stddev=stddev; dst=.Normal; mean=0f32} {elem_type=float; dim=weight.basic.dim}

            inb a = s.CudaBlas.trmm .Left .Lower .nT .NonUnit 1f32 front.sampling.basic random.basic |> CudaAux.temporary
            inb b = s.CudaBlas.trmm .Right .Lower .nT .NonUnit 1f32 back.sampling.basic a |> CudaAux.temporary

            //{ d with weight = View.zip {(View.unzip weight) with primal = (s.CudaBlas.geam' .nT .nT one self.basic one b.basic random; View.wrap weight.dim random)} }
            d
        | _ ->
            d
```

As a sanity check for whether weight sampling is being done too much, let me turn this on."