"6:05pm.
```
        apply = inl {weights={weights outer} input} s ->
            inl weights = weight_sample weights s // TODO: Make this stateful even for FF nets.
            inl apply =
                inm data = concat {bias=one; input}
                matmult_stream {weights with data} >>= activation
            apply s
```

Agh, this. This is the culprit. This is also the reason why the net is so unstable. Let me fix this.

```
        apply = inl {d with weights={weights outer} input} s ->
            inl weights =
                match d with
                | {state={weights}} -> weights
                | _ -> weight_sample weights s
            inl apply =
                inm data = concat {bias=one; input}
                matmult_stream {weights with data} >>= activation
            inl {out bck} = apply s
            {out state={weights}; bck}
```

Let me try it like this.

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:06.9651269 for: iteration 0
Training: 4.85559929039001
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:00.0263359 for: iteration 1
Training: NaN
Training diverged. Aborting...
```

Ah, now it is even more unstable. Let me lower the learning rate."