"6:10pm.
```
{rate = {covariance = 2.478797E-05; weight = 3.814697E-06}}
The learning rate is 2 ** -18
Starting timing for: iteration 0
The time was 00:00:06.9562560 for: iteration 0
Training: 4.85559929039001
The learning rate is 2 ** -18
Starting timing for: iteration 1
The time was 00:00:06.7714432 for: iteration 1
Training: 4.27563108618864
The learning rate is 2 ** -18
Starting timing for: iteration 2
The time was 00:00:06.7991283 for: iteration 2
Training: 4.35683184363605
The learning rate is 2 ** -18
Starting timing for: iteration 3
The time was 00:00:06.7548925 for: iteration 3
Training: 4.46776081466883
The learning rate is 2 ** -18
Starting timing for: iteration 4
The time was 00:00:06.7838626 for: iteration 4
Training: 8.67709079949076
```

With weight sampling it is just so bad. Not at all what I expected.

Let me gradually turn down the noise. I want to make sure that nothing strange is going on."