``` {rate = {covariance = 0.0004716485; weight = 0.0001220703}} The learning rate is 2 ** -13 Starting timing for: iteration 0 The time was 00:00:07.1880935 for: iteration 0 Training: 2.27501921951497 The learning rate is 2 ** -13 Starting timing for: iteration 1 The time was 00:00:07.1190833 for: iteration 1 Training: 1.87663735857453 The learning rate is 2 ** -13 Starting timing for: iteration 2 The time was 00:00:07.1000626 for: iteration 2 Training: 1.79421502233976 The learning rate is 2 ** -13 Starting timing for: iteration 3 The time was 00:00:07.1247800 for: iteration 3 Training: 1.74990055538193 The learning rate is 2 ** -13 Starting timing for: iteration 4 The time was 00:00:07.1092247 for: iteration 4 Training: 1.72193165255628 ```
Yeah, this is what I would expect from noisy KFAC. Of course it won't be good as the regular thing during training since I am just measuring over-fitting here, but it is quite close. The important thing is that the net is no longer going haywire. I can in fact take this to RL land.

Tomorrow, I am going to look into speeding up the the compilation - right now the slowness has reached critical levels and I need to do some digging here.

I also need to investigate why is there such a huge difference in training cost on the first epoch compared to last time. As a note, I did change the activation from `tanh` to `ln_relu`, but that should not be the cause.

In addition to that, why did the training slow down from 5.2 to 6.9s for the RNN? No doubt this is somehow connected to why I am suddenly getting better performance, but I am not sure what is going on exactly here.

I should also clean up `truncate`.

Well, I am done with this bit of overtime. I am glad I could end the day on a proper note.