I got up over half an hour ago, but needed to take some time to answer a question. As it is getting close to break time, I won't start right away.
Today my mission will be investigating the large slowdown in compilation in the recent version of the library.

To be honest, I am not sure I will be able to fix the slowdown, but I'll give it my best shot. Last time I fretted because of it, but it turned out that join points everywhere were the solution.

This time, the join points already are present in most of the places, so comparatively speaking the slowdown cannot be for the same reason. But I'll defer judgement until I give it a try.