```     inl network =         ln_relu 512,         ln_relu 512,         ln_relu 512,         ln_relu 512,         ln_relu 512,         linear label_size ```
When I do this it climbs to 3.25. Let me add a bit more.

```
    inl network =
        ln_relu 512,
        ln_relu 512,
        ln_relu 512,
        ln_relu 512,
        ln_relu 512,
        ln_relu 512,
        ln_relu 512,
        linear label_size
```

This is 3.54. Yeah, it is not good scaling, but not quite exponential. No doubt there will be ways to wring something out.

```
    inl network =
        linear label_size
```

How about this?

This is 1.11s.

```
    inl network =
        ln_relu 512,
        linear label_size
```

This is 2s. Yeah, no doubt too much inlining is happening.