0.6s, 2.15k LOC.
Hmm, this is a bit big, but still not too much. I mean, it is a lot, but within the range I expected. A lot of boilerplate is needed to set things up.

Let me take a look at the produced code. Then I will comment in the lines one by one.