```     __global__ void method_31(long long int var_0, long long int var_1, float var_2, float * __restrict__ var_3, long long int var_4, long long int var_5);     __global__ void method_39(long long int var_0, long long int var_1, float var_2, float * __restrict__ var_3, long long int var_4, long long int var_5);     __global__ void method_45(long long int var_0, long long int var_1, float * __restrict__ var_2, long long int var_3, long long int var_4, float * __restrict__ var_5, long long int var_6, long long int var_7);     __device__ char method_32(long long int var_0, long long int * __restrict__ var_1); ```
Now the amount of kernels is quite a bit less. I lowered the compilation time by a few % and the amount of code from 2.1k to 1.9k.

I am still not done with the initializer. I'll put join points everywhere that I can.

```
| .init -> Struct.iter2' (met (!dyn init) (!dyn tns) -> init tns s) init (View.unzip tns)
```

Let me start with this.

Now the compilation time feel to 0.55s, but the amount of code is 2k.

```
            | .init ->
                inl tns = dyn tns
                inl init = dyn init
                join Struct.iter2' (met init tns -> init tns s) init (View.unzip tns)
```

What about this?

Now the compilation is 0.55s and the amount of code is 1.93k LOC.

Yeah, this one is best. Since I just have one layer I can't really feel the effect of my changes, but it should be more pronounced for multiple layers.

```
            met stddev_sum_init (!dyn mult) (!dyn tns) =
                inl stddev = sqrt (mult / to float32 (Tuple.foldl (+) 0 tns.dim))
                inl mean = 0f32
                normal {stddev mean} tns
```

Let me also dyn this thing...No it is fine.

Now...

Let me comment in all the stuff that I removed so I can observe the overall effect of these changes.